Add files via upload
we have a Wireless Sensor Network (WSN). It includes a controller and several
sensors. The same as the previous assignment, these sensors collect environment parameters
and send them to the only controller. However, there are two malicious nodes that are hackers by
a network attacker. The attacker uses one of them to implement the UDP Flood attack and uses
the other one to implement the Ping Flood attack. Please capture the network packets, analyze
the data and classify the network packets with machine learning algorithms.