Add files via upload
Assume that we are standing at an intersection. There are a total of four directions
of traffic passing the intersection. In order to maintain the smooth flow of traffic, we need four
signal lights to indicate the passing traffic. When the red light is on, vehicles should be prohibited
from passing. When the green light is on, vehicles can pass.